fix: enhance promotion display status with priority and exclusivity labels
